requirements.txt
gunicorn
cherrypy
google-cloud-storage
PyYAML
